Springdata arreglado (no funciona el crud)

+++ src/main/java/com/org/iesalixar/daw2/OrtegaAlvaro/dwese_ticket_logger_webapp/entities/Province.java
@@ -18,7 +18,7 @@ public class Province {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @NotEmpty(message = "{msg.province.code.notEmpty}")
     @Size(max = 10, message = "{msg.province.code.size}")
@@ -46,7 +46,7 @@ public class Province {
         this.region = region;
     }
 
-    public int getRegionId() {
+    public Long getRegionId() {
         return region != null ? region.getId() : 0;
     }
 }
